chore: move language to dedicated package
Independent of editor
