+++ lib/editor/index.js
@@ -6,7 +6,7 @@ import { EditorState } from '@codemirror/state';
 import { EditorView, keymap, tooltips, lineNumbers } from '@codemirror/view';
 
 import { parser as markdownParser } from '@lezer/markdown';
-import { createFeelersLanguageSupport } from './language/index.js';
+import { createFeelersLanguageSupport } from '../lang/index.js';
 
 import { feelersLinter } from './lint/index.js';
 import { lightTheme, darkTheme } from '@bpmn-io/cm-theme';
