+++ lib/lang/index.js
