+++ lib/lang/mixedParser.js
@@ -1,5 +1,5 @@
 import { parser as feelParser } from '@bpmn-io/lezer-feel';
-import { parser as templateParser } from '../../parser/parser.js';
+import { parser as templateParser } from '../parser/parser.js';
 import { parseMixed } from '@lezer/common';
 import { LRLanguage, foldNodeProp, foldInside } from '@codemirror/language';
